+++ CHANGELOG.md
@@ -26,6 +26,8 @@ The goal for the 0.7 series is to support an internal plugin model. There should
 - `SDConfig` instance is now a global in `plugin_utils`. Despite the oft-repeated advice not to use globals, it really seems to make sense in this context. It simplifies most plugin utility function signatures. For example: `plugin_utils.write_output(self.sd_config, msg)` is now `plugin_utils.write_output(msg)`. There are numerous ways to protect the config instance if the need arises, or choose a different non-globabl approach with this as a better starting point.
 - `SDConfig.validate()` checks that attributes required by all plugins have been set.
 - Plugins import `sd_config` directly. This simplifies the use of `sd_config` significantly in plugins. For example, `self.sd_config.automate_all` becomes `sd_config.automate_all`. Also, core no longer needs to pass anything off directly to plugins.
+- `SDConfig` is instantiated once, at the module level, in `plugin_utils`. It's imported directly into simple_deploy.py and each plugin's `platform_deployer` module. Attributes are set by `Command`, and read by plugins as needed.
+- `SimpleDeployCommandError` is moved to its own module, and imported directly into any module that needs it.
 
 
 ### 0.7.2